requirements.txt
pip==24.2
transformers==4.42.4
langchain_core==0.2.21
torch==2.2.2
uuid==1.30
gtts==2.5.2
sounddevice==0.4.7
setuptools==72.1.0
wave==0.0.2
SpeechRecognition==3.10.4
pyaudio==0.2.14
python==3.10.14
streamlit==1.38.0
tts==0.22.0
ffmpeg==6.1.1
pydub==0.25.1
ipykernel==6.29.5